Refactor listable classes namespace
This commit refactors listable classes namepsace.
